+++ coders/mat.c
@@ -179,7 +179,7 @@ register PixelPacket *q;
     if(*p>0)
       {
       f=(*p/Max)*(MaxRGB-q->red);
-      if((int) f+q->red>MaxRGB) q->red=MaxRGB;
+      if(f+q->red>MaxRGB) q->red=MaxRGB;
                      else q->red+=(int) f;
             if((int) f/2.0>q->green) q->green=q->blue=0;
          else q->green=q->blue-=(int) (f/2.0);    
@@ -187,7 +187,7 @@ register PixelPacket *q;
     if(*p<0)
       {
       f=(*p/Max)*(MaxRGB-q->blue);
-      if((int) f+q->blue>MaxRGB) q->blue=MaxRGB;
+      if(f+q->blue>MaxRGB) q->blue=MaxRGB;
                       else q->blue+=(int) f;
             if((int) f/2.0>q->green) q->green=q->red=0;
         else q->green=q->red-=(int) (f/2.0);
